Fix il2cpp static guards and range checks for long-based slot collections

+++ Scellecs.Morpeh/Core/Collections/LongSlotMap.cs
@@ -35,7 +35,7 @@
         
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool Has(long key) {
-            var rem = key & this.capacityMinusOne;
+            var rem = (int)(key & this.capacityMinusOne);
 
             int next;
             for (var i = this.buckets[rem] - 1; i >= 0; i = next) {
@@ -52,7 +52,7 @@
         
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool Remove(long key, out int slotIndex) {
-            var rem = key & this.capacityMinusOne;
+            var rem = (int)(key & this.capacityMinusOne);
 
             int next;
             int num = -1;
@@ -93,7 +93,7 @@
         
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool TryGetIndex(long key, out int slotIndex) {
-            var rem = key & this.capacityMinusOne;
+            var rem = (int)(key & this.capacityMinusOne);
 
             int next;
             for (var i = this.buckets[rem] - 1; i >= 0; i = next) {
@@ -112,7 +112,7 @@
         
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool IsKeySet(long key, out int slotIndex) {
-            var rem = key & this.capacityMinusOne;
+            var rem = (int)(key & this.capacityMinusOne);
 
             for (var i = this.buckets[rem] - 1; i >= 0; i = this.slots[i].next) {
                 if (this.slots[i].key - 1 == key) {
@@ -154,7 +154,7 @@
                 ++this.lastIndex;
             }
             
-            var rem = key & this.capacityMinusOne;
+            var rem = (int)(key & this.capacityMinusOne);
             ref var newSlot = ref this.slots[slotIndex];
 
             newSlot.key = key + 1;
